+++ src/memory_stream.h
@@ -26,9 +26,9 @@ class MemoryStream : public StreamInterface {
   // the |memory|.
   static UniqueStreamPtr CreateForWrite(Buffer* memory);
 
-  bool GetSize(size_t* size) const override;
-  bool GetOffset(size_t* offset) const override;
-  bool Seek(size_t offset) override;
+  bool GetSize(uint64_t* size) const override;
+  bool GetOffset(uint64_t* offset) const override;
+  bool Seek(uint64_t offset) override;
   bool Read(void* buffer, size_t length) override;
   bool Write(const void* buffer, size_t length) override;
   bool Close() override;
@@ -44,7 +44,7 @@ class MemoryStream : public StreamInterface {
   Buffer* write_memory_;
 
   // The current offset.
-  size_t offset_;
+  uint64_t offset_;
 
   // True if the stream is open.
   bool open_;
